requirements.txt
# Core libraries
streamlit
chromadb
sentence-transformers
langchain
beautifulsoup4
requests
google-genai

--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.8.0+cpu